added source

+++ proseminar/09/naive-nqueens.c
@@ -14,7 +14,7 @@ https://stackoverflow.com/questions/18661702/passing-matrix-as-a-parameter-in-fu
 #include <string.h>
 
 // Total amount of possible solutions
-int size;
+unsigned long long size;
 
 /* A utility function to print solution */
 void printSolution(int N, int board[N][N]) {
@@ -106,10 +106,10 @@ bool solveNQ(int N) {
     solveNQUtil(0, N, board);
     double end_time = omp_get_wtime();
     if (size == 0) {
-        printf("%d, %f\n", size, end_time - start_time);
+        printf("%lld, %f\n", size, end_time - start_time);
         return false;
     } else {
-        printf("%d, %f\n", size, end_time - start_time);
+        printf("%lld, %f\n", size, end_time - start_time);
         return true;
     }
 }
